Fixed issues #3 and #4

+++ src/qpack/qpack.c
@@ -1355,7 +1355,7 @@ static qp_types_t QP_print_unpacker(
     switch (tp)
     {
     case QP_INT64:
-        printf("%ld", qp_obj->via.int64);
+        printf("%lld", (long long) qp_obj->via.int64);
         break;
     case QP_DOUBLE:
         printf("%f", qp_obj->via.real);
