+++ src/siri/db/shards.c
@@ -230,8 +230,8 @@ int siridb_shards_add_points(
                         pend)) < 0)
                 {
                     log_critical(
-                            "Could not write points to shard id '%ld",
-                            shard->id);
+                            "Could not write points to shard id '%llu",
+                            (unsigned long long) shard->id);
                 }
                 else
                 {
